make CPU templates vendor specific
