+++ src/cpuid/src/template/mod.rs
@@ -1,7 +1,5 @@
 // Copyright 2019 Amazon.com, Inc. or its affiliates. All Rights Reserved.
 // SPDX-License-Identifier: Apache-2.0
 
-/// Follows a C3 template in setting up the CPUID.
-pub mod c3;
-/// Follows a T2 template in setting up the CPUID.
-pub mod t2;
+// Contains Intel specific templates.
+pub mod intel;
